requirements.txt
absl-py==0.13.0
anyascii==0.2.0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blis==0.4.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.6
chardet==4.0.0
contractions==0.0.52
cymem==2.0.5
debugpy==1.3.0
decorator==5.0.9
deep-translator==1.4.4
defusedxml==0.7.1
detectlanguage==1.5.0
Django==2.2.5
django-cors-headers==3.8.0
djangorestframework==3.12.4
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.2.5/en_core_web_lg-2.2.5.tar.gz
entrypoints==0.3
flatbuffers==1.12
ftfy==6.0.3
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==2.10
importlib-metadata==3.10.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.2.5
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plac==1.1.3
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.7.7
ptyprocess==0.7.0
pyahocorasick==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
spacy==2.2.4
spotipy==2.18.0
sqlparse==0.4.1
srsly==1.0.5
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
textsearch==0.0.21
thinc==7.4.0
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==6.2.0
wrapt==1.12.1
xgboost==0.90
zipp==3.5.0
